move area to another test suite
